Adds a new contact type, bbInt, to ConFind. Incorporates backbone interaction and interference into fasstDB, allowing these to be stored along with normal contacts. Finally, adds an option to the ConFind constructor that allows it to tolerate residues missing backbone atoms.
ConFind::bbInteraction(resA, resB) returns the distance of the closest
backbone atoms between the two residues. ignoreFlanking controls whether
residues flanking the residue of interest are considered when reporting
bbInteractions. For convenience, other functions are provided to match
the CD/Int functions.

FasstDB is modified to incorporate interference and bb interaction. As
interference is directional, it is stored as TWO residue pair
properties: interfering and interfered. The former returns the residues
that interfere with residue of interest and the latter returns the
residues that the residue of interest interferes with.