Update dtermen.cpp
If the number of matches from the current clique ends up being less than selfCorrMinN, that just means that one or more matches were eliminated from the solution set due to being too similar to the query (this happens in dTERMen::termData::setMatches). So this, in and of itself, does not _technically_ meet the criterion for the clique not to grow. The fix I propose is to simply look at the last solution, whether it is the solution with index (selfCorrMinN - 1) or perhaps one or two indices earlier.