Update dtermen.cpp
Reverting back to Sebastian's original fix to the dTERMen::selfEnergies bug. The two solutions are in practice going to work the same way (under all reasonable choices of number of matches and cutoffs), but the original solution seems more stable and reliable, because technically it is possible to end up with zero matches after termData::setMatches (not likely but formally possible), so explicitly checking whether the remaining number of matches reaches selfCorrMinN seems better.